refactor: move stagelist methods to game class
